cleanup(drivers): use helper methods
