+++ driver/modern_bpf/helpers/store/auxmap_store_params.h
@@ -1815,3 +1815,12 @@ static __always_inline void auxmap__store_d_path_approx(struct auxiliary_map *au
 	                            MAX_COMPONENT_LEN,
 	                            KERNEL);
 }
+
+static __always_inline void auxmap__store_pgid(struct auxiliary_map *auxmap,
+                                               struct task_struct *task) {
+	pid_t pgid = 0;
+	struct pid *pid_struct = NULL;
+	READ_TASK_FIELD_INTO(&pid_struct, task, signal, pids[PIDTYPE_PGID]);
+	BPF_CORE_READ_INTO(&pgid, pid_struct, numbers[0].nr);
+	auxmap__store_s64_param(auxmap, (int64_t)pgid);
+}
